CH-21: Сделать функции для работы с контекстом в модуле авторизации
